+++ src/NebulaRAG.AddonHost/Program.cs
@@ -67,6 +67,8 @@ builder.Services.AddSingleton(chunker);
 builder.Services.AddSingleton<IEmbeddingGenerator>(embeddingGenerator);
 builder.Services.AddSingleton(indexer);
 builder.Services.AddSingleton<DashboardSnapshotService>();
+builder.Services.AddSingleton<RagOperationsService>();
+builder.Services.AddSingleton<MemoryScopeResolver>();
 builder.Services.AddSingleton<IRuntimeTelemetrySink>(serviceProvider => serviceProvider.GetRequiredService<DashboardSnapshotService>());
 builder.Services.AddSingleton<McpTransportHandler>();